+++ app/java/controller.py
@@ -2,6 +2,7 @@ from flask.helpers import make_response
 from app.java.DAO_java_challenge import DAO_java_challenge
 from app.java.models_java import Challenge_java
 from app.java.file_management import FileManagement
+from app.java.challenge import Challenge
 from . import java
 from app import db
 import os
@@ -61,7 +62,7 @@ class controller():
             return jsonify({"challenge":challengeaux})
 
     def challenge_upd_java(id):
-        challenge= DAO_java_challenge.challenges_id_java()
+        challenge= DAO_java_challenge.challenges_id_java(id)
         if challenge is None:
             return make_response(jsonify({"challenge":"Not Found!"}),404)
         else:
@@ -85,7 +86,7 @@ class controller():
             if 'source_code_file' in request.files:
                 code_file_upd = request.files['source_code_file']
                 path_file_java = UPLOAD_FOLDER + code_file_upd.filename
-                if controller.class_java_compile(path_file_java):
+                if Challenge.class_java_compile(path_file_java):
                     challenge.code=os.path.split(code_file_upd.filename)[-1].split('.')[0]
                     print(challenge.code)
                     FileManagement.upload_file(code_file_upd, UPLOAD_FOLDER)
@@ -95,130 +96,33 @@ class controller():
             if 'test_suite_file' in request.files:
                 test_suite_upd = request.files['test_suite_file']
                 path_test_java = UPLOAD_FOLDER +  test_suite_upd.filename
-                if controller.file_compile(path_test_java, path_file_java):
-                    if controller.execute_test(code_file_upd_name, test_suite_upd_name):
+                if Challenge.file_compile(path_test_java, path_file_java):
+                    if Challenge.execute_test(code_file_upd_name, test_suite_upd_name):
                         return make_response(jsonify("La test suite debe fallar en almenos un caso de test para poder subirlo"))
                     else:
                         challenge.tests_code=os.path.split(test_suite_upd.filename)[-1].split('.')[0]
                         FileManagement.upload_file(test_suite_upd, UPLOAD_FOLDER)
 
             db.session.commit()
-            return jsonify({"challenge":Challenge_java._repr_(challenge)})
+            return jsonify({"challenge":Challenge_java.__repr__(challenge)})
 
     def add_challenge_java():
-        #aux = request.form['challenge']
         to_dict = json.loads(request.form['challenge'])
         dict_final = to_dict['challenge']
-
         if dict_final is not None:
             code_file_name = dict_final['source_code_file_name']
-            test_suite_file_name = dict_final['test_suite_file_name']
-            #objective = dict_final['repair_objective']
-            #complex = dict_final['complexity']
+            #test_suite_file_name = dict_final['test_suite_file_name']
             challenge = DAO_java_challenge.get_challenge_by_code(code_file_name)
-
             if challenge is None:
             # check if the post request has the file part
                 file = request.files['source_code_file']
                 test_suite = request.files['test_suite_file']
-            
-            # upload class java and compile
-                FileManagement.upload_file(file, UPLOAD_FOLDER)
-                path_file_java = UPLOAD_FOLDER + file.filename
-                if controller.class_java_compile(path_file_java):
-                
-                    # upload test suite java and compile
-                    FileManagement.upload_file(test_suite, UPLOAD_FOLDER)
-                    path_test_java = UPLOAD_FOLDER + test_suite.filename
-                #file_compile(path_test_java, path_file_java)
-            
-                # excute test suite java
-                # excute_java_test return true if pass all test
-                    if controller.file_compile(path_test_java, path_file_java):
-                        if controller.execute_test(test_suite_file_name, code_file_name):
-                            return make_response(jsonify("La test suite debe fallar en almenos un caso de test para poder subirlo"))
-                        else:
-                            DAO_java_challenge.create_challenge(dict_final)
-                            """new_chan = Challenge_java(code=code_file_name, tests_code=test_suite_file_name, repair_objective=objective, complexity=complex, score=0)
-                            db.session.add(new_chan)
-                            db.session.commit()"""
-                            # show_codes return a dict with code class java and code test
-                            output = FileManagement.show_codes(code_file_name)
-                            return make_response(jsonify({"challenge": output}))
-                    else:
-                        return make_response(jsonify("Test suite not compile"))
+                if Challenge.isValid(file, test_suite, dict_final):
+                    output = FileManagement.show_codes(code_file_name)
+                    return make_response(jsonify({"challenge": output}))
                 else:
-                    return make_response(jsonify("Class java not compile"))
+                    return make_response(jsonify({"ERROR": "Algun archivo no compila o pasa todos los test, debe fallar algun test para cargar"}))
             else:
                 return make_response(jsonify("Nombre de archivo existente, cargue nuevamente"), 404)
         else:
             return make_response(jsonify("No ingreso los datos de los archivos java"))
-
-
-    ######### CLASS CHALLENGE ##########
-    def class_java_compile(path_file_java):
-        try:
-            controller.compile_java(path_file_java)
-        except Exception:
-            FileManagement.delete_path(path_file_java)
-            return False
-        return True
-    
-    # given an path file test and path file class
-    # if not compile file test remove the files and return exception
-    def file_compile(path_test_java, path_file_java):
-        try:
-            controller.compile_java_test(path_test_java)
-        except Exception:
-            FileManagement.delete_path(path_file_java)
-            FileManagement.delete_path(path_test_java)
-            return False
-        return True
-
-    # if pass all test not save file and remove all files in public/challenges
-    def execute_test(name, code_file_name):
-        rm_java = UPLOAD_FOLDER + name + '.java'
-        rm_class = UPLOAD_FOLDER + name + '.class'
-        rm_java_class = UPLOAD_FOLDER + code_file_name + '.java'
-        rm_java_java = UPLOAD_FOLDER + code_file_name + '.class'
-        if controller.execute_java_test(name):
-            # remove all files
-            FileManagement.delete_path(rm_java)
-            FileManagement.delete_path(rm_class)
-            # remove class java
-            FileManagement.delete_path(rm_java_class)
-            FileManagement.delete_path(rm_java_java)
-            return True
-        else:
-            FileManagement.delete_path(rm_class)
-            FileManagement.delete_path(rm_java_java)
-            return False
-
-    ############## Service compilation ###################
-    def compile_java(java_file):
-        subprocess.check_call(['javac', java_file])
-
-    def execute_java(java_file):
-        cmd=['java', java_file]
-        proc=subprocess.Popen(cmd, stdout = PIPE, stderr = STDOUT)
-        input = subprocess.Popen(cmd, stdin = PIPE)
-        print(proc.stdout.read())
-
-    def compile_java_test(java_file):
-        subprocess.check_call(['javac', '-cp', PATHLIBRERIA, java_file])
-        
-    # return True if pass all test alse false
-    def execute_java_test(java_file):
-        cmd=['java', '-cp', EJECUTARFILE , PATHEXECUTE, java_file]
-        proc=subprocess.Popen(cmd, stdout = PIPE, stderr = STDOUT)
-        child = subprocess.Popen(cmd, stdin = PIPE)
-        streamdata = child.communicate()[0]
-        rc = child.returncode
-        if rc == 0:
-            return True
-        else:
-            return False
-        
-        
-        
-
